perf: reduce spectrogram memory 50x (byte[] dB instead of double[], 300x256 instead of 600x512), lazy bitmap building, batch UI progress every 10 files

+++ LosslessChecker/Models/AnalysisResult.cs
@@ -33,5 +33,5 @@ public record AnalysisResult
     public AnalysisStatus AnalysisStatus { get; init; } = AnalysisStatus.Pending;
 
     public double[] AveragedSpectrum { get; init; } = Array.Empty<double>();
-    public double[][] SpectrogramData { get; init; } = Array.Empty<double[]>();
+    public byte[][] SpectrogramData { get; init; } = Array.Empty<byte[]>();
 }
